style(WPN1): the eleven lints clippy -D warnings found in this wave's tests
Ten `.clone()` calls on `CrowdArchetype`, which is `Copy` -- the fixtures in
`weapon_3d`, `phase30_gameplay_gate`, `crowd` and `witness` were all cloning a
value that copies -- and two `as f32` casts on `SpriteInstance::size`, which
is already `f32` (the neighbouring `position` casts are real, which is why the
pattern looked right).

All eleven are in test code this wave added; nothing in the shipped crates
moved. `cargo clippy --workspace --all-targets` with `RUSTFLAGS=-D warnings`
is exit 0 with zero errors and zero warnings.

+++ crates/inf-ui/src/view.rs
@@ -426,7 +426,7 @@ mod tests {
         let mut below = 0;
         for q in &l.quads {
             let (x, y) = (q.position.x as f32, q.position.y as f32);
-            let (w, h) = (q.size.x as f32, q.size.y as f32);
+            let (w, h) = (q.size.x, q.size.y);
             // Nothing covers the middle.
             assert!(
                 !(x <= cx && cx <= x + w && y <= cy && cy <= y + h),
